add voice-chat-util to make play audio commands more generic
