+++ src/commands/audio/kys.js
@@ -1,28 +1,20 @@
 const { SlashCommandBuilder } = require('discord.js');
-const ytdl = require('@distube/ytdl-core');
-const { joinVoiceChannel, createAudioPlayer, createAudioResource } = require('@discordjs/voice');
+const { playAudioInVoiceChannel } = require('../../utils/voice-chat-util');
 
 module.exports = {
         data: new SlashCommandBuilder()
                 .setName('kys')
                 .setDescription('young man'),
         async execute(interaction) {
-
                 const url = 'https://www.youtube.com/watch?v=SQsmW0mMKZc'
-                const stream = ytdl(url, { filter: 'audioonly', highWaterMark: 1 << 25, });
-                const player = createAudioPlayer();
-                const resource = createAudioResource(stream);
-                const guildMember = await interaction.member.guild.members.fetch(interaction.user.id);
-                const { channelId } = guildMember.voice;
-                const connection = joinVoiceChannel({
-                        channelId: channelId,
-                        guildId: interaction.guildId,
-                        adapterCreator: interaction.channel.guild.voiceAdapterCreator,
-                });
-                connection.subscribe(player);
-                player.play(resource);
-                await interaction.reply("young man");
 
+                try {
+                        await playAudioInVoiceChannel(interaction, url);
 
+                        await interaction.reply("young man");
+                } catch (error) {
+                        console.error('Error:', error);
+                        await interaction.reply('An error occurred while trying to play the audio.');
+                }
         },
 };
