+++ src/commands/audio/ponder.js
@@ -1,6 +1,5 @@
 const { SlashCommandBuilder } = require('discord.js');
-const ytdl = require('@distube/ytdl-core');
-const { joinVoiceChannel, createAudioPlayer, createAudioResource } = require('@discordjs/voice');
+const { playAudioInVoiceChannel } = require('../../utils/voice-chat-util');
 
 module.exports = {
         data: new SlashCommandBuilder()
@@ -10,24 +9,7 @@ module.exports = {
                 const url = 'https://www.youtube.com/watch?v=AXqMnPyx73E';
 
                 try {
-                        const stream = ytdl(url, { filter: 'audioonly', highWaterMark: 1 << 25 });
-                        const resource = createAudioResource(stream);
-                        const guildMember = await interaction.member.guild.members.fetch(interaction.user.id);
-                        const { channelId } = guildMember.voice;
-
-                        if (!channelId) {
-                                return interaction.reply('You must join a voice channel first!');
-                        }
-
-                        const connection = joinVoiceChannel({
-                                channelId: channelId,
-                                guildId: interaction.guildId,
-                                adapterCreator: interaction.channel.guild.voiceAdapterCreator,
-                        });
-
-                        const player = createAudioPlayer();
-                        player.play(resource);
-                        connection.subscribe(player);
+                        await playAudioInVoiceChannel(interaction, url);
 
                         await interaction.reply("brain brain brain brain");
                 } catch (error) {
